refactor(payments): retire Batch 2 payment proxy files and align SSOT imports
