+++ backend/api/src/__tests__/services/PaymentService.spec.ts
@@ -15,13 +15,13 @@ import {
     createPaymentTransaction, 
     getUserForPayment 
 } from '@esparex/core/services/TransactionService';
-import { getPlanById } from '@esparex/core/services/PlanService';
+import { getPlanById } from '@esparex/core/domains/payments/application/PlanService';
 import { getRazorpayClient, getRazorpayRuntimeConfig } from '@esparex/core/config/razorpay';
 
 // ── Mocks ────────────────────────────────────────────────────────────────────
 
 jest.mock('@esparex/core/services/TransactionService');
-jest.mock('@esparex/core/services/PlanService');
+jest.mock('@esparex/core/domains/payments/application/PlanService');
 jest.mock('@esparex/core/config/razorpay');
 jest.mock('@esparex/core/utils/logger');
 jest.mock('../../utils/errorResponse', () => ({
