--- core/src/services/PlanEngine.ts
@@ -1,3 +1,0 @@
-// @deprecated
-// Remove during M6 or M8
-export { calculateUserPlan } from "../domains/payments";
